+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/ContentDto.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.data.remote.dto.response
+
+data class ContentDto(
+    val id: Int,
+    val quantity: Int,
+    val product: ProductDto,
+)
