+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/ProductResponse.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.shopping.dto
+package woowacourse.shopping.data.remote.dto.response
 
 data class ProductResponse(
     val content: List<ProductDto>,
