+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/SortDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.shopping.dto
+package woowacourse.shopping.data.remote.dto.response
 
 data class SortDto(
     val sorted: Boolean,
