+++ app/src/main/java/woowacourse/shopping/data/remote/shopping/ProductApiService.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data.remote.shopping
 
 import retrofit2.Call
-import woowacourse.shopping.data.remote.shopping.dto.ProductDto
-import woowacourse.shopping.data.remote.shopping.dto.ProductResponse
+import woowacourse.shopping.data.remote.dto.response.ProductDto
+import woowacourse.shopping.data.remote.dto.response.ProductResponse
 
 interface ProductApiService {
     fun loadById(productId: Long): Call<ProductDto>
